track_loader_old/requirements.txt
beautifulsoup4
requests
google-cloud-storage
google-cloud-firestore